requirements.txt
streamlit>=1.28
torch>=2.0
timm>=0.9
pillow>=10.0
pandas>=2.0
matplotlib>=3.7
seaborn>=0.12
numpy>=1.24